+++ kairo-health-check/feature/src/main/kotlin/kairo/healthCheck/HealthCheckRep.kt
@@ -1,4 +1,4 @@
-package kairo.healthCheck.feature
+package kairo.healthCheck
 
 import kotlinx.serialization.Serializable
